Add confirmation to /scamban & improve /activate
Added ConfirmBanView which displays information about the soon to be banned user in question.

Also added deferred responses to /activate

Added the ability for /scamcheck to show shared servers that a scammer may be in. This is via the config "ScamCheckShowsSharedServers" which defaults to true.

This option requires both the presence and members intent.

Closes #30